cloud/requirements.txt
# Web 服务框架
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# 数据校验
pydantic>=2.0.0

# 图片处理
Pillow>=10.0.0

# Qwen2.5-VL 模型推理（仅 server_qwen.py 需要）
torch>=2.1.0
transformers>=4.49.0
accelerate>=0.30.0
bitsandbytes>=0.43.0

# qwen_vl_utils 为本地源码包，不通过 pip 安装
# 路径：./Qwen/qwen-vl-utils/src
# 已在 server_qwen.py 启动时自动加入 sys.path，无需额外操作